+++ tests/unit/test_model.py
@@ -574,7 +574,7 @@ class TestFromPreprocessorBinary:
 
         with (
             patch(
-                "pyiwfm.io.preprocessor_binary.PreprocessorBinaryReader.read",
+                "pyiwfm.io.binary.preprocessor.PreprocessorBinaryReader.read",
                 return_value=mock_data,
             ) as mock_read,
             patch(
@@ -601,7 +601,7 @@ class TestFromPreprocessorBinary:
 
         with (
             patch(
-                "pyiwfm.io.preprocessor_binary.PreprocessorBinaryReader.read",
+                "pyiwfm.io.binary.preprocessor.PreprocessorBinaryReader.read",
                 return_value=mock_data,
             ),
             patch(
@@ -626,7 +626,7 @@ class TestFromPreprocessorBinary:
 
         with (
             patch(
-                "pyiwfm.io.preprocessor_binary.PreprocessorBinaryReader.read",
+                "pyiwfm.io.binary.preprocessor.PreprocessorBinaryReader.read",
                 return_value=mock_data,
             ),
             patch(